Abstracted Classes
Note that the dependencies have to be rebuilt along with the runs. I could only manage to get the run to work properly on 'runClient' as the custom client task didn't have the mixin parameter passed in. Might be something to look at.

The australian creeper needs more information to correctly fix the model.